openrisc: Enable stopping debug unit on all traps
The original code only enabled traps if sw breakpoints were enabled, but
this means we wont break on single stepping.  Enable so single stepping
and software breakpoints work.

This means we can remoke an unused util function too.

Also, it mentions the kernel relies on l.trap instructions, but thats not
longer true as far as I can see.

Change-Id: If62193f5382717d5f8b67b8aeb5100035d0046e8
Signed-off-by: Stafford Horne <shorne@gmail.com>